requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
configparser==5.0.1
cycler==0.10.0
docker-pycreds==0.4.0
gitdb==4.0.5
GitPython==3.1.9
h5py==3.1.0
idna==2.10
imbalanced-learn==0.7.0
joblib==0.17.0
kiwisolver==1.2.0
llvmlite==0.35.0
matplotlib==3.3.2
numba==0.52.0
numpy==1.19.2
pandas==1.1.3
pathtools==0.1.2
patsy==0.5.1
Pillow==8.0.0
plotly==4.11.0
promise==2.3
protobuf==3.13.0
psutil==5.7.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
retrying==1.3.3
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.0
sentry-sdk==0.18.0
shap==0.37.0
shortuuid==1.0.1
six==1.15.0
slicer==0.0.3
smmap==3.0.4
statsmodels==0.12.1
subprocess32==3.5.4
threadpoolctl==2.1.0
tqdm==4.50.2
urllib3==1.25.10
wandb==0.10.7
watchdog==0.10.3
xgboost==1.2.1
